Battleship game | v. 0.2
This version allows you to play with a randomly generated grid,
You can set your grid and based on that grid the randomly generated grid will generate!

SO for example: if you set a grid with 3 ships with sizes (3,2,4)
THEN, computer will generate a grid with these 3 ships in random direction and in random position.

Explaining each function would take a lot of time so I won't, but take a look if you want. I will explain all of it later (hopefully).

NEXT UPDATES:

- Computer playing with you,
- Online game,
- Fixes.