Online Game | v. 0.5
This update introduces awaited online gameplay (by who? - by no one).
This update allows you to connect with other people and play with them.

You have to choose option '1', then choose if you want to be the "host" or listen for host to start the game.
The connection is opened using IP which host writes.

IMPORTANT: like any other application for this to work you need to allow the application through firewall :+1:

"host" is the person starting game, in programming he sends the TCP message.
"client" is the person listening for message from the host, he also sends his messages too (obviously)

The host chooses how many ships will be in the game, because he is the one who sets the grid. Host sets the grid and then the program calculates amount of ships, then client sets his grid based on hosts grid.

After that they play and have fun :)!

NOTE: This version is proven to contain bugs and issues, I know! I will fix it in next update :)
This version is playable but you can easily break it :(

This is the v. 0.5, because it introduced Computer playing with you, Online game, and small fixes.

NEXT UPDATES:

- Fixing issues.
- Refactoring,
- Who knows (Hmm?)