Refactor and Fixes | v. 0.7
Fixes is v. 0.6,
Refactor v. 0.7

Fixed some errors, and introduced a better way of counting amount of ships and their respective sizes on a grid.

Refactorizes the code, making it more developer-friendly and possibly increasing its performance. Deleted obsolete and duplicate code.

This update is nothing fancy, it's more of a maintenance thing.

NEXT UPDATES:
- For now unknown